Support templates in output file name
This eases the work with complex filenames or aids the creation of
scripts utilizing stretch where the input file may not be known prior.
Limited path changes are also possible.

As it makes little sense to encode multiple files to the same target
location, it also paves the way for batch processing.

This technique and in fact nearly the whole commit are stolen from a
commit of the same title in the hardsub project [1].

[1]: https://github.com/dffischer/hardsub/commit/bc5a46a1853becdfe1ba4d511995c8be38e181c8