Accept arbitrary factors
The prior limitation stemmed from the atempo filter which can only
handle values in the interval in question. Stacking multiple instances
of the filter, however, amends this problem. The technique is taken from
the same site that sparked the idea to the stretch tool [1].

The preparation is not integrated inside the argument parsing loop as
the -s option may occur multiple times. Although this does not make much
sense in human-composed command line, a script may utilize this. By
moving the composing step after argument parsing, it is always only
executed once.

[1]: https://trac.ffmpeg.org/wiki/How%20to%20speed%20up%20/%20slow%20down%20a%20video