Permit decimals with implicit zero integral part
Most programs and programming languages allow omitting a single zero
left of the decimal dot, so users may be used to using this format.
Scripts may also feed such numbers to stretch. As all programs the
number in passed on to can handle it, there is no point in restricting
its use.